AI-Career-Companion/requirements.txt
streamlit
google-generativeai
pydantic
pytest
psycopg2-binary
redis
sqlalchemy
pdfplumber
python-docx